backend/requirements.txt
# --- Core Math & JIT ---
numpy==1.26.4
llvmlite==0.42.0
numba==0.59.1
pymatting==1.1.12

# --- Web Framework & Async ---
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9
aiofiles==23.2.1
python-dotenv==1.0.1

# --- Video / Image Processing ---
opencv-python-headless==4.9.0.80
Pillow==10.3.0
imageio==2.34.1
imageio-ffmpeg==0.5.1
mediapipe>=0.10.13

# --- Background Removal & ML ---
rembg==2.0.57
onnxruntime==1.18.0
scipy==1.13.1
scikit-image==0.24.0
pooch==1.8.1
jsonschema==4.23.0

# --- Progress & Utilities ---
tqdm==4.66.4

pydantic-settings==2.2.1